+++ frontend/src/components/EntityAdd.vue
@@ -33,6 +33,7 @@ Modal window to add alliances or corporations to the database.
 </template>
 
 <script>
+import {toRefs} from "vue";
 import _ from 'lodash';
 import {Modal} from "bootstrap";
 import Multiselect from '@suadelabs/vue3-multiselect';
@@ -44,13 +45,12 @@ export default {
         Multiselect,
     },
 
-    props: {
-        settings: Object,
-    },
+    inject: ['store'],
 
     data: function() {
         return {
             h: new Helper(this),
+            settings: toRefs(this.store.state).settings,
             addType: '', // alliance or corp
             searchIsLoading: false,
             searchResults: [],