+++ frontend/src/pages/RoleAdmin.vue
@@ -47,7 +47,6 @@ export default {
     },
 
     props: {
-        settings: Object,
         route: Array,
         player: Object,
     },